Boxing/unboxing clean-up

+++ southbound/southbound-impl/src/main/java/org/opendaylight/ovsdb/southbound/transactions/md/OvsdbNodeRemoveCommand.java
@@ -91,7 +91,7 @@ public class OvsdbNodeRemoveCommand extends AbstractTransactionCommand {
         This is to handle the controller initiated connection scenario, where all the controller will connect, but
         switch will have only one manager.
         */
-        if (onlyConnectedManager.getNumberOfConnections().longValue() > ONE_ACTIVE_CONNECTION_IN_PASSIVE_MODE) {
+        if (onlyConnectedManager.getNumberOfConnections() > ONE_ACTIVE_CONNECTION_IN_PASSIVE_MODE) {
             return false;
         }
         return true;
